[#74] TokenRepository로 이름 수정
